Project retirement notice
As stated in the readme, I'll be putting all keyboard projects on indefinite hiatus and releasing them to the public domain. All future updates will be canceled, and absolutely no support will be provided. All the files are provided as-is, exactly where they've stopped being developed.

No firmware will be provided.